7 years later...
Updated dependencies, and managed to start the gui.
A shitty gui, btw. Not that I ever thought the UX
was nice, but I didn't remember it as that bad.

Also removed a bunch of pointless comments. Style
was a requirement by the course this project belonged to.

Some functionality probably broke, because I had to
get rid of proprietary libs.

Ok. See you in a few years...